show relationship between clusters and pathways

Hierarchical clustering of clusters and pathways according to GSEA result. If by = "GO" and pathwayIDs = NULL, only GO terms of level 5-6 are used for calculation.

hierarchyplot_tree(
  SeuratObj,
  by = "GO",
  pathwayIDs = NULL,
  topaths = 5,
  cluster_cutree_k = NULL,
  pathway_cutree_k = NULL,
  color.use.cluster = NULL,
  color.use.pathway = NULL,
  weight.scale = TRUE,
  vertex.size.cex = 1,
  vertex.label.cex = 0.9,
  edge.max.width = 3,
  vertex.label.color = "black",
  alpha.edge = 0.6
)

Arguments

SeuratObj
Object of class "Seurat"

by
which GSEA result to show

pathwayIDs
IDs of pathways to show

topaths
top n pathway of each cluster

cluster_cutree_k
the clusters are divided into k, based on the hierarchical clustering (using cutree)

pathway_cutree_k
the pathways are divided into k, based on the hierarchical clustering (using cutree)

color.use.cluster
used to color the cluster nodes, a named vector or a vector

color.use.pathway
used to color the pathway nodes

weight.scale
scale the width or not

vertex.size.cex
node size

vertex.label.cex
size of node label

edge.max.width
width of edge

vertex.label.color
color of label of nodes, 'black' by default

alpha.edge
transparency of edge color

Examples

if (FALSE) {
hierarchyplot_tree(SeuratObj, by = "GO", topaths = 10, cluster_cutree_k = 5, pathway_cutree_k = 20)
}